% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gretlR-package.R
\docType{package}
\name{gretlR-package}
\alias{gretlR}
\alias{gretlR-package}
\title{gretlR: A Seamless Integration of 'Gretl' and 'R'}
\description{
It allows running 'gretl' (<http://gretl.sourceforge.net/index.html>) program from R, R Markdown and Quarto. 'gretl' ('Gnu' Regression, 'Econometrics', and Time-series Library) is a statistical software for Econometric analysis. This package does not only integrate 'gretl' and 'R' but also serves as a 'gretl' Knit-Engine for 'knitr' package. Write all your 'gretl' commands in 'R', R Markdown chunk.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=gretlR}
  \item Report bugs at \url{https://github.com/sagirumati/gretlR/issues}
}


Other important functions: 
\code{\link{eng_gretl}()},
\code{\link{exec_gretl}()},
\code{\link{exec_inp}()},
\code{\link{import_kable1}()},
\code{\link{import_kable}()},
\code{\link{include_graph}()},
\code{\link{include_tex}()},
\code{\link{write_inp}()}
}
\author{
\strong{Maintainer}: Sagiru Mati \email{smati@smati.com.ng} (\href{https://orcid.org/0000-0003-1413-3974}{ORCID})

}
\concept{important functions}
\keyword{documentation}
